requirements.txt
aiogram
python-dotenv
aiohttp
aiohttp-socks
telethon
beautifulsoup4
lxml